Refactor player animation
Player animation is now contained within the player class and called by the map function. Road generation and collision detection have been refactored to use for loops.
