+++ bin/.gitignore
@@ -11,3 +11,4 @@
 /Building$Status.class
 /Warehouse.class
 /Vehicle.class
+/Player.class
